src/chatbot_backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pinned to 4.0.x to avoid passlib compatibility warning

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.1

# Utils
python-multipart==0.0.9
python-dotenv==1.0.1
